Restructure pts.js storybook
